refactor: properly organize tests - unit vs integration
CRITICAL FIX: Unit tests must NEVER touch the filesystem!

Changes:
- Moved filesystem-dependent tests from unit tests to integration tests
- Removed redundant `integration/` subfolder from tests directory
- Renamed test files to remove redundant `_integration` suffix
- Moved tests from generic integration_tests.rs to appropriate unit tests:
  - test_serialization_roundtrip -> types/session.rs unit tests
  - test_full_type_integration -> types/session.rs unit tests
- Fixed duplicate RollingLogConfig definition (use from config.rs)
- Fixed all clippy warnings about format strings
- Updated Cargo.toml to reference correct test paths

Test organization now follows Rust conventions:
- Unit tests: In #[cfg(test)] modules within source files
- Integration tests: In tests/ directory (which use filesystem/external resources)

This fixes the testing architecture violations where unit tests were
incorrectly using TempDir and filesystem operations.
