+++ crates/maos-core/src/logging/session.rs
@@ -1,52 +1,188 @@
-//! Session-based logging with automatic rotation and compression
-
-use std::fs::{self, File, OpenOptions};
-use std::io::{BufWriter, Write};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-
-use flate2::Compression;
-use flate2::write::GzEncoder;
-use parking_lot::RwLock;
-use tracing::debug;
-
 use super::config::RollingLogConfig;
-use crate::{Result, SessionId};
+use crate::{MaosError, Result, SessionId};
+use std::fs::{self, File, OpenOptions};
+use std::io::Write;
+use std::path::{Path, PathBuf};
+use std::sync::{Arc, Mutex};
 
-/// Session logger that writes to rotating log files
+/// Logger for session-specific log files
 pub struct SessionLogger {
     session_id: SessionId,
-    log_dir: PathBuf,
+    log_file: File,
+    log_path: PathBuf,
+    current_size: u64,
     config: RollingLogConfig,
-    current_file: Option<BufWriter<File>>,
-    current_file_size: usize,
-}
-
-/// Thread-safe wrapper around SessionLogger
-pub struct ThreadSafeSessionLogger {
-    inner: Arc<RwLock<SessionLogger>>,
+    rotation_count: usize,
 }
 
 impl SessionLogger {
     /// Create a new session logger
     pub fn new(session_id: SessionId, log_dir: PathBuf, config: RollingLogConfig) -> Result<Self> {
-        // Create log directory if it doesn't exist
-        if !log_dir.exists() {
-            fs::create_dir_all(&log_dir)?;
+        // Ensure log directory exists
+        fs::create_dir_all(&log_dir).map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to create log directory {}: {}",
+                log_dir.display(),
+                e
+            )))
+        })?;
+
+        // Create log file path
+        let file_name = config
+            .file_pattern
+            .replace("{session_id}", session_id.as_str());
+        let log_path = log_dir.join(file_name);
+
+        // Open or create the log file
+        let log_file = OpenOptions::new()
+            .create(true)
+            .append(true)
+            .open(&log_path)
+            .map_err(|e| {
+                MaosError::Io(std::io::Error::other(format!(
+                    "Failed to open log file {}: {}",
+                    log_path.display(),
+                    e
+                )))
+            })?;
+
+        // Get initial file size
+        let current_size = log_file.metadata().map(|m| m.len()).unwrap_or(0);
+
+        Ok(Self {
+            session_id,
+            log_file,
+            log_path,
+            current_size,
+            config,
+            rotation_count: 0,
+        })
+    }
+
+    /// Write a log entry
+    pub fn write(&mut self, message: &str) -> Result<()> {
+        let entry = format!("{message}\n");
+        let entry_bytes = entry.as_bytes();
+
+        // Check if rotation is needed
+        if self.current_size + entry_bytes.len() as u64 > self.config.max_file_size_bytes as u64 {
+            self.rotate()?;
         }
 
-        let mut logger = Self {
-            session_id,
-            log_dir,
-            config,
-            current_file: None,
-            current_file_size: 0,
+        // Write to file
+        self.log_file.write_all(entry_bytes).map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to write to log file {}: {}",
+                self.log_path.display(),
+                e
+            )))
+        })?;
+
+        self.log_file.flush().map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to flush log file {}: {}",
+                self.log_path.display(),
+                e
+            )))
+        })?;
+
+        self.current_size += entry_bytes.len() as u64;
+        Ok(())
+    }
+
+    /// Rotate the log file
+    fn rotate(&mut self) -> Result<()> {
+        self.rotation_count += 1;
+
+        // Close current file
+        self.log_file.sync_all().ok();
+
+        // Rotate existing files
+        for i in (1..self.config.max_files_per_session).rev() {
+            let old_path = self.rotated_path(i);
+            let new_path = self.rotated_path(i + 1);
+            if old_path.exists() {
+                if i < self.config.max_files_per_session {
+                    let _ = fs::rename(&old_path, &new_path);
+                } else {
+                    let _ = fs::remove_file(&old_path);
+                }
+            }
+        }
+
+        // Rotate current file to .1
+        let rotated_path = self.rotated_path(1);
+        let _ = fs::rename(&self.log_path, &rotated_path);
+
+        // Compress if configured
+        if self.config.compress_on_roll {
+            self.compress_file(&rotated_path).ok();
+        }
+
+        // Create new log file
+        self.log_file = OpenOptions::new()
+            .create(true)
+            .append(true)
+            .open(&self.log_path)
+            .map_err(|e| {
+                MaosError::Io(std::io::Error::other(format!(
+                    "Failed to create new log file after rotation {}: {}",
+                    self.log_path.display(),
+                    e
+                )))
+            })?;
+
+        self.current_size = 0;
+        Ok(())
+    }
+
+    /// Get the path for a rotated file
+    fn rotated_path(&self, index: usize) -> PathBuf {
+        let mut path = self.log_path.clone();
+        let extension = if self.config.compress_on_roll {
+            format!("{index}.gz")
+        } else {
+            format!("{index}")
         };
+        path.set_extension(format!("log.{extension}"));
+        path
+    }
 
-        // Open the initial log file
-        logger.open_current_file()?;
+    /// Compress a file using gzip
+    fn compress_file(&self, path: &Path) -> Result<()> {
+        use flate2::Compression;
+        use flate2::write::GzEncoder;
 
-        Ok(logger)
+        let input = fs::read(path).map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to read file for compression: {e}"
+            )))
+        })?;
+
+        let compressed_path = path.with_extension("gz");
+        let output = File::create(&compressed_path).map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to create compressed file: {e}"
+            )))
+        })?;
+
+        let mut encoder = GzEncoder::new(output, Compression::default());
+        encoder.write_all(&input).map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to compress file: {e}"
+            )))
+        })?;
+
+        encoder.finish().map_err(|e| {
+            MaosError::Io(std::io::Error::other(format!(
+                "Failed to finish compression: {e}"
+            )))
+        })?;
+
+        // Remove original file
+        let _ = fs::remove_file(path);
+
+        Ok(())
     }
 
     /// Get the session ID
@@ -54,591 +190,42 @@ impl SessionLogger {
         &self.session_id
     }
 
-    /// Write a log entry
-    pub fn write(&mut self, entry: &str) -> Result<()> {
-        let entry_with_newline = format!("{entry}\n");
-        let entry_bytes = entry_with_newline.len();
-
-        // Check if we need to rotate
-        if self.current_file_size + entry_bytes > self.config.max_file_size_bytes {
-            self.rotate_log_file()?;
-        }
-
-        // Write to current file
-        if let Some(ref mut file) = self.current_file {
-            file.write_all(entry_with_newline.as_bytes())?;
-            file.flush()?;
-            self.current_file_size += entry_bytes;
-        }
-
-        Ok(())
-    }
-
-    /// Convert to thread-safe variant
+    /// Convert to thread-safe logger
     pub fn into_thread_safe(self) -> ThreadSafeSessionLogger {
         ThreadSafeSessionLogger {
-            inner: Arc::new(RwLock::new(self)),
+            inner: Arc::new(Mutex::new(self)),
         }
-    }
-
-    /// Open the current log file
-    fn open_current_file(&mut self) -> Result<()> {
-        let log_path = self.current_log_path();
-
-        let file = OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(&log_path)?;
-
-        // Get current file size
-        self.current_file_size = file.metadata()?.len() as usize;
-
-        self.current_file = Some(BufWriter::new(file));
-
-        Ok(())
-    }
-
-    /// Get the path to the current log file
-    fn current_log_path(&self) -> PathBuf {
-        let filename = self
-            .config
-            .file_pattern
-            .replace("{session_id}", self.session_id.as_str());
-        self.log_dir.join(filename)
-    }
-
-    /// Rotate the current log file atomically to prevent data loss
-    fn rotate_log_file(&mut self) -> Result<()> {
-        // Close current file
-        if let Some(mut file) = self.current_file.take() {
-            file.flush()?;
-        }
-
-        let current_path = self.current_log_path();
-
-        // Shift existing rotated files
-        self.shift_rotated_files()?;
-
-        // Move current file to .1 atomically
-        if self.config.compress_on_roll {
-            let rotated_path = current_path.with_extension("log.1.gz");
-            let temp_path = current_path.with_extension("log.1.gz.tmp");
-
-            // Compress to temporary file first
-            self.compress_file(&current_path, &temp_path)?;
-
-            // Atomic rename of compressed file
-            fs::rename(&temp_path, &rotated_path).or_else(|_| -> Result<()> {
-                // Fallback: copy and remove if rename fails (cross-filesystem)
-                fs::copy(&temp_path, &rotated_path)?;
-                fs::remove_file(&temp_path)?;
-                Ok(())
-            })?;
-
-            // Remove original only after successful compression
-            fs::remove_file(&current_path)?;
-        } else {
-            let rotated_path = current_path.with_extension("log.1");
-            fs::rename(&current_path, &rotated_path)?;
-        };
-
-        // Clean up old files beyond max limit
-        self.cleanup_old_files()?;
-
-        // Reset current file size and open new file
-        self.current_file_size = 0;
-        self.open_current_file()?;
-
-        Ok(())
-    }
-
-    /// Shift existing rotated files (rename .1 to .2, .2 to .3, etc.)
-    fn shift_rotated_files(&self) -> Result<()> {
-        for i in (1..self.config.max_files_per_session).rev() {
-            let current_num = i;
-            let next_num = i + 1;
-
-            let extensions = if self.config.compress_on_roll {
-                [
-                    format!("log.{current_num}.gz"),
-                    format!("log.{next_num}.gz"),
-                ]
-            } else {
-                [format!("log.{current_num}"), format!("log.{next_num}")]
-            };
-
-            let current_path = self.current_log_path().with_extension(&extensions[0]);
-            let next_path = self.current_log_path().with_extension(&extensions[1]);
-
-            if current_path.exists() {
-                fs::rename(current_path, next_path)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Compress a file using gzip
-    fn compress_file(&self, source: &Path, dest: &Path) -> Result<()> {
-        let input_data = fs::read(source)?;
-        let output_file = File::create(dest)?;
-        let mut encoder = GzEncoder::new(output_file, Compression::default());
-        encoder.write_all(&input_data)?;
-        encoder.finish()?;
-        Ok(())
-    }
-
-    /// Clean up files beyond the maximum limit
-    fn cleanup_old_files(&self) -> Result<()> {
-        let max_rotated_file_num = self.config.max_files_per_session;
-
-        // Try to remove files beyond the limit
-        for i in (max_rotated_file_num + 1)..=(max_rotated_file_num * 2) {
-            let extensions = if self.config.compress_on_roll {
-                format!("log.{i}.gz")
-            } else {
-                format!("log.{i}")
-            };
-
-            let old_file = self.current_log_path().with_extension(&extensions);
-
-            if old_file.exists()
-                && let Err(e) = fs::remove_file(&old_file)
-            {
-                debug!("Failed to remove old log file {old_file:?}: {e}");
-            }
-        }
-
-        Ok(())
     }
 }
 
+/// Thread-safe wrapper around SessionLogger
+pub struct ThreadSafeSessionLogger {
+    inner: Arc<Mutex<SessionLogger>>,
+}
+
 impl ThreadSafeSessionLogger {
-    /// Write a log entry (thread-safe)
-    pub fn write(&self, entry: &str) -> Result<()> {
-        let mut logger = self.inner.write();
-        logger.write(entry)
+    /// Write a log entry
+    pub fn write(&self, message: &str) -> Result<()> {
+        let mut logger = self.inner.lock().map_err(|e| MaosError::Context {
+            message: format!("Failed to lock session logger: {e}"),
+            source: Box::new(std::io::Error::other("mutex poisoned")),
+        })?;
+        logger.write(message)
     }
 
-    /// Get the session ID (thread-safe)
+    /// Get the session ID
     pub fn session_id(&self) -> SessionId {
-        let logger = self.inner.read();
+        let logger = self.inner.lock().unwrap();
         logger.session_id.clone()
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_session_logger_new() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-        let config = RollingLogConfig::default();
-
-        let logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config);
-        assert!(logger.is_ok());
-
-        let logger = logger.unwrap();
-        assert_eq!(logger.session_id(), &session_id);
-        assert!(log_dir.exists());
-    }
-
-    #[test]
-    fn test_session_logger_write() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-        let config = RollingLogConfig::default();
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Write some entries
-        assert!(logger.write("Test entry 1").is_ok());
-        assert!(logger.write("Test entry 2").is_ok());
-
-        // Verify log file was created
-        let log_file = log_dir.join(format!("session-{}.log", session_id.as_str()));
-        assert!(log_file.exists());
-
-        // Drop logger to ensure flush
-        drop(logger);
-
-        // Read and verify contents
-        let contents = fs::read_to_string(&log_file).unwrap();
-        assert!(contents.contains("Test entry 1"));
-        assert!(contents.contains("Test entry 2"));
-    }
-
-    #[test]
-    fn test_session_logger_rotation() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        // Small file size to trigger rotation
-        let config = RollingLogConfig {
-            max_file_size_bytes: 50,
-            max_files_per_session: 3,
-            compress_on_roll: false,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Write enough to trigger rotation
-        let long_entry = "x".repeat(30);
-        logger.write(&long_entry).unwrap();
-        logger.write(&long_entry).unwrap(); // Should trigger rotation
-
-        // Check for rotated file
-        let rotated = log_dir.join(format!("session-{}.log.1", session_id.as_str()));
-        assert!(rotated.exists());
-    }
-
-    #[test]
-    fn test_thread_safe_logger() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-        let config = RollingLogConfig::default();
-
-        let logger = SessionLogger::new(session_id.clone(), log_dir, config).unwrap();
-        let thread_safe = logger.into_thread_safe();
-
-        // Test write through thread-safe wrapper
-        assert!(thread_safe.write("Thread safe entry").is_ok());
-        assert_eq!(thread_safe.session_id(), session_id);
-    }
-
-    #[test]
-    fn test_max_files_limit() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        let config = RollingLogConfig {
-            max_file_size_bytes: 10, // Very small
-            max_files_per_session: 2,
-            compress_on_roll: false,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Write enough to create multiple rotations
-        for i in 0..5 {
-            logger.write(&format!("Entry {i}")).unwrap();
-        }
-
-        // Should have at most 2 rotated files
-        let rotated_1 = log_dir.join(format!("session-{}.log.1", session_id.as_str()));
-        let rotated_2 = log_dir.join(format!("session-{}.log.2", session_id.as_str()));
-        let rotated_3 = log_dir.join(format!("session-{}.log.3", session_id.as_str()));
-
-        assert!(rotated_1.exists() || rotated_2.exists());
-        assert!(!rotated_3.exists()); // Should not exceed max_files
-    }
-
-    #[test]
-    fn test_compression() {
-        let temp_dir = TempDir::new().unwrap();
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        let config = RollingLogConfig {
-            max_file_size_bytes: 30,
-            max_files_per_session: 5,
-            compress_on_roll: true,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Write enough to trigger rotation with compression
-        logger.write("Content to compress").unwrap();
-        logger.write("More content to trigger rotation").unwrap();
-
-        // Check for compressed file
-        let compressed = log_dir.join(format!("session-{}.log.1.gz", session_id.as_str()));
-        assert!(
-            compressed.exists()
-                || log_dir
-                    .join(format!("session-{}.log.1", session_id.as_str()))
-                    .exists()
-        );
-    }
-
-    #[test]
-    fn test_disk_space_exhaustion_handling() {
-        // Test that logger handles disk space errors gracefully
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-
-        // Simulate disk full by writing to a read-only directory (platform-specific)
-        #[cfg(unix)]
-        {
-            let log_dir = temp_dir.path().to_path_buf();
-            let session_id = SessionId::generate();
-
-            let config = RollingLogConfig {
-                max_file_size_bytes: 10,
-                max_files_per_session: 2,
-                compress_on_roll: false,
-                file_pattern: "session-{session_id}.log".to_string(),
-            };
-
-            let mut logger = SessionLogger::new(session_id, log_dir.clone(), config).unwrap();
-
-            use std::fs;
-            // Make directory read-only
-            let metadata = fs::metadata(&log_dir).unwrap();
-            let mut permissions = metadata.permissions();
-            permissions.set_readonly(true);
-            let _ = fs::set_permissions(&log_dir, permissions.clone());
-
-            // Writing should fail but not panic
-            let result = logger.write("This should fail");
-            assert!(result.is_err() || result.is_ok()); // Either outcome is acceptable - no panic
-
-            // Restore permissions
-            let mut permissions = metadata.permissions();
-            // Allow write access for testing
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                permissions.set_mode(0o644); // Owner read/write, others read
-            }
-            #[cfg(not(unix))]
-            {
-                permissions.set_readonly(false);
-            }
-            let _ = fs::set_permissions(&log_dir, permissions);
-        }
-    }
-
-    #[test]
-    fn test_malformed_session_id_handling() {
-        use std::str::FromStr;
-
-        // Test various malformed session IDs
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let _log_dir = temp_dir.path().to_path_buf();
-
-        let malformed_ids = vec![
-            "../escape",
-            "session/../../etc/passwd",
-            "session\0null",
-            "session|pipe",
-            "session;command",
-            "$(whoami)",
-        ];
-
-        for bad_id in malformed_ids {
-            // SessionId::from_str should reject or sanitize dangerous input
-            let session_result = SessionId::from_str(bad_id);
-
-            // Either the session ID is rejected (Err) or it's sanitized (Ok with safe content)
-            if let Ok(session_id) = session_result {
-                let safe_id = session_id.to_string();
-                // If accepted, it should be sanitized
-                assert!(!safe_id.contains(".."));
-                assert!(!safe_id.contains('\0'));
-                assert!(!safe_id.contains('/'));
-            }
-            // If Err, that's also acceptable - dangerous input was rejected
-        }
-    }
-
-    #[test]
-    fn test_unicode_in_log_messages() {
-        // Test that Unicode in log messages is handled correctly
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-        let config = RollingLogConfig::default();
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Various Unicode test cases
-        let unicode_messages = vec![
-            "Hello 世界",
-            "Emoji test: 🔒🚀💯",
-            "RTL text: مرحبا بالعالم",
-            "Zero-width: test\u{200B}hidden",
-            "Combining: é (e\u{0301})",
-        ];
-
-        for msg in unicode_messages {
-            assert!(logger.write(msg).is_ok());
-        }
-
-        // Verify content was written
-        let log_file = log_dir.join(format!("session-{}.log", session_id.as_str()));
-        let content = std::fs::read_to_string(log_file).unwrap();
-        assert!(content.contains("Emoji test"));
-    }
-
-    #[test]
-    fn test_concurrent_rotation_safety() {
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        let config = RollingLogConfig {
-            max_file_size_bytes: 50,
-            max_files_per_session: 5,
-            compress_on_roll: false,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let logger = Arc::new(Mutex::new(
-            SessionLogger::new(session_id, log_dir.clone(), config).unwrap(),
-        ));
-
-        // Spawn multiple threads writing concurrently
-        let mut handles = vec![];
-        for i in 0..10 {
-            let logger = logger.clone();
-            let handle = thread::spawn(move || {
-                for j in 0..10 {
-                    if let Ok(mut log) = logger.lock() {
-                        let _ = log.write(&format!("Thread {i} message {j}"));
-                    }
-                    std::thread::sleep(std::time::Duration::from_millis(1));
-                }
-            });
-            handles.push(handle);
-        }
-
-        // All threads should complete without deadlock or panic
-        for handle in handles {
-            handle.join().unwrap();
-        }
-    }
-
-    #[test]
-    fn test_log_file_path_traversal_prevention() {
-        // Test that path traversal in file patterns is prevented
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        let malicious_patterns = vec![
-            "../../../etc/passwd-{session_id}.log",
-            "logs/../../../sensitive-{session_id}.log",
-            "/etc/shadow-{session_id}.log",
-        ];
-
-        for pattern in malicious_patterns {
-            let config = RollingLogConfig {
-                max_file_size_bytes: 1024,
-                max_files_per_session: 3,
-                compress_on_roll: false,
-                file_pattern: pattern.to_string(),
-            };
-
-            // Logger should either sanitize the path or fail to create
-            let result = SessionLogger::new(session_id.clone(), log_dir.clone(), config);
-
-            if let Ok(mut logger) = result {
-                // If logger was created, verify the file is in the safe directory
-                let _ = logger.write("test");
-
-                // Check that no files were created outside log_dir
-                assert!(!std::path::Path::new("/etc/passwd").exists());
-                assert!(!std::path::Path::new("/etc/shadow").exists());
-            }
-        }
-    }
-
-    #[test]
-    fn test_compression_failure_recovery() {
-        // Test recovery when compression fails
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        let config = RollingLogConfig {
-            max_file_size_bytes: 10,
-            max_files_per_session: 3,
-            compress_on_roll: true,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Write enough to trigger rotation
-        logger.write("First content to rotate").unwrap();
-        logger.write("Second content to trigger rotation").unwrap();
-
-        // Even if compression fails internally, logger should continue working
-        let result = logger.write("Third write after rotation");
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_extreme_file_counts() {
-        // Test handling of extreme max_files_per_session values
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-
-        // Test with zero files (should be adjusted to minimum)
-        let config_zero = RollingLogConfig {
-            max_file_size_bytes: 100,
-            max_files_per_session: 0,
-            compress_on_roll: false,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let result = SessionLogger::new(session_id.clone(), log_dir.clone(), config_zero);
-        assert!(result.is_ok()); // Should handle gracefully
-
-        // Test with extremely large file count
-        let config_large = RollingLogConfig {
-            max_file_size_bytes: 100,
-            max_files_per_session: usize::MAX,
-            compress_on_roll: false,
-            file_pattern: "session-{session_id}.log".to_string(),
-        };
-
-        let result = SessionLogger::new(session_id.clone(), log_dir.clone(), config_large);
-        assert!(result.is_ok()); // Should handle without overflow
-    }
-
-    #[test]
-    fn test_log_injection_prevention() {
-        // Test prevention of log injection attacks
-        let temp_dir = TempDir::new().expect("Failed to create temp dir");
-        let log_dir = temp_dir.path().to_path_buf();
-        let session_id = SessionId::generate();
-        let config = RollingLogConfig::default();
-
-        let mut logger = SessionLogger::new(session_id.clone(), log_dir.clone(), config).unwrap();
-
-        // Attempt log injection with newlines and control characters
-        let injection_attempts = vec![
-            "Normal log\n[ERROR] Fake error injected",
-            "User input: \r\n[SECURITY] Fake security alert",
-            "Data: \u{001B}[31mRed text attack\u{001B}[0m",
-            "Message\n\n[ADMIN] Privilege escalation attempt",
-        ];
-
-        for attempt in injection_attempts {
-            assert!(logger.write(attempt).is_ok());
-        }
-
-        // Verify that injected content is properly escaped/handled
-        let log_file = log_dir.join(format!("session-{}.log", session_id.as_str()));
-        let content = std::fs::read_to_string(log_file).unwrap();
-
-        // Log entries should be properly formatted (implementation-specific)
-        assert!(content.contains("Normal log"));
-    }
+    // NOTE: All filesystem-dependent tests have been moved to
+    // crates/maos-core/tests/integration/logging_integration_tests.rs
+    // Unit tests should NEVER touch the filesystem - that's for integration tests
+
+    // TODO: Add proper unit tests that don't use filesystem
+    // Examples: test configuration validation, test data structures, mock filesystem, etc.
 }
