+++ crates/maos-core/src/types/session.rs
@@ -91,6 +91,8 @@ pub enum SessionStatus {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::types::agent::{AgentId, AgentInfo, AgentStatus};
+    use crate::types::tool::{ToolCall, ToolCallId};
     use std::path::PathBuf;
 
     #[test]
@@ -178,5 +180,56 @@ mod tests {
             let deserialized: SessionStatus = serde_json::from_str(&json).unwrap();
             assert_eq!(status, deserialized);
         }
+    }
+
+    #[test]
+    fn test_session_id_serialization_roundtrip() {
+        let session_id = SessionId::generate();
+
+        // Serialize and deserialize
+        let json = serde_json::to_string(&session_id).unwrap();
+        let deserialized: SessionId = serde_json::from_str(&json).unwrap();
+
+        assert_eq!(session_id, deserialized);
+        assert!(deserialized.is_valid());
+    }
+
+    #[test]
+    fn test_full_type_integration() {
+        // Create a session
+        let session = Session {
+            id: SessionId::generate(),
+            created_at: Utc::now(),
+            last_activity: Utc::now(),
+            status: SessionStatus::Active,
+            workspace_root: PathBuf::from("/tmp/maos-test"),
+            active_agents: vec![],
+        };
+
+        // Create an agent
+        let agent = AgentInfo {
+            id: AgentId::generate(),
+            agent_type: "test-agent".to_string(),
+            session_id: session.id.clone(),
+            workspace_path: PathBuf::from("/tmp/maos-test/agent-workspace"),
+            status: AgentStatus::Active,
+            created_at: Utc::now(),
+            last_activity: Utc::now(),
+        };
+
+        // Create a tool call
+        let tool_call = ToolCall {
+            id: ToolCallId::generate(),
+            tool_name: "Bash".to_string(),
+            parameters: serde_json::json!({ "command": "echo test" }),
+            timestamp: Utc::now(),
+            session_id: Some(session.id.clone()),
+            agent_id: Some(agent.id.clone()),
+        };
+
+        // Verify everything works together
+        assert!(session.id.is_valid());
+        assert!(agent.id.is_valid());
+        assert_eq!(tool_call.session_id.as_ref().unwrap(), &session.id);
     }
 }
