+++ crates/maos-core/tests/invalid_paths_tests.rs
